Create W0ne7_Arduino_Uno_Example.ino
This is a simple arduino script for use with PTRB's W0ne7 (seven way switch). This code will result in any direction press or rotation of the switch outputting a keystroke to the connected PC. 

In doing so, this can be used to map keyboard presses in games/simulators to the W0ne7.